// src/types/preknow.ts
export interface Attachment {
  id: number;
  original: string;
  thumbnail: string;
  __typename?: string;
  slug?: string;
}
// ...
export interface User {
  _id: string;
  name: string;
  email: string;
  contact?: string;
  image: string;
  bio?: string;
  delivery_address?: string;
  wallet: {
    total_points: number;
    points_used: number;
    available_points: number;
  };
  profile: {
    id?: string;
    contact?: string;
    bio?: string;
    avatar?: Attachment;
  };
  address: string;
}

export interface UpdateUserInput extends Partial<User> {
  _id: string;
}